requirements.txt
requests>=2.31.0
python-dotenv==1.0.0
mypy==1.8.0
flake8==7.0.0
httpx==0.26.0
pydantic==2.6.1
pytest==8.0.0
pytest-cov==4.1.0
isort==5.13.2
black==24.1.1